+++ modules/core/include/rmvl/core.hpp
@@ -43,7 +43,6 @@
 #include "core/numcal.hpp"
 
 // 数据、信号处理
-#include "core/ew_topsis.hpp"
 #include "core/kalman.hpp"
 
 // 数据结构
